+++ tests/0040-io_event.c
@@ -141,7 +141,7 @@ int main_0040_io_event (int argc, char **argv) {
 #ifndef _MSC_VER
 			r = read(pfd.fd, &b, 1);
 #else
-                        r = _read(pfd.fd, &b, 1);
+                        r = _read((int)pfd.fd, &b, 1);
 #endif
 			if (r == -1)
 				TEST_FAIL("read failed: %s\n", strerror(errno));
